Create C Program
You are in charge of the cake for a child's birthday. You have decided the cake will have one candle for each year of their total age. They will only be able to blow out the tallest of the candles. Count how many candles are tallest.

Candle heights are [3,2,1,3]. The tallest candles are 3 units, and there are 2 of them.